Score close-readiness; improve desktop shutdown
Add server-side close-readiness scoring and component breakdown so UIs can render a scored readiness payload (new OperationsCloseReadinessComponentDto, Score and Components on close DTO, helper to build components and blockers). Update contract blocker lists and direct-lending error codes. Implement WPF BackfillViewModel activation lifecycle with cancellation/dispose handling and update view/code-behind to use Activate/Deactivate. Improve dev desktop launcher to generate and inject a per-run shutdown token, request graceful host shutdown via POST /api/system/shutdown, wait for exit, and safely stop the owned WPF process if needed. Misc: fix Aggressor cast to byte, clean up endpoint DI/signatures and tests to match DTO/behavior changes, and update READMEs and tests accordingly.

+++ src/Meridian.Application/Pipeline/DualPathEventPipeline.cs
@@ -304,7 +304,7 @@ public sealed class DualPathEventPipeline : IMarketEventPublisher, IBackpressure
                 symbolHash,
                 trade.Price,
                 trade.Size,
-                (int)trade.Aggressor,
+                (byte)trade.Aggressor,
                 evt.Sequence);
 
             if (_tradeBuffer.TryWrite(in raw))
